pyproject.toml
[tool.poetry]
name = "django-educa"
version = "0.1.0"
description = ""
authors = ["macalistervadim <macalistervadim@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.5"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
django-cleanup = "^9.0.0"
django-cors-headers = "^4.6.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
mypy = "^1.14.1"
django-stubs = "^5.1.1"
ruff = "^0.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    "*/migrations/*",
    "venv/",
    "__init__.py",
    "apps.py",
]
line-length = 79
indent-width = 4

[tool.ruff.lint]
select = ["E", "W", "C", "F", "N", "COM", "I", "Q", "U", "N", "A", "M"]
ignore = ["F405"]
fixable = ["ALL"]
unfixable = []
